refactor(engine): extract WorkspaceBuilder + Warmup + banned-token list
Combines three small init-tail methods into one TU:
  - Engine::init_features (~77 LOC)
  - Engine::build_banned_token_list (~85 LOC)
  - Engine::warmup (~91 LOC)

All three run during engine init after weights + KV cache are set up;
colocating them keeps the init-tail logic in one file. Declarations in
engine.h unchanged. Bodies byte-identical.

Phase 4 of docs/superpowers/specs/2026-05-20-architecture-refactor-roadmap-design.md
